Add support for ACPI 6.4 and the 64-bit XDsdt offset
`XDsdt` contains the 64-bit address of the DSDT in memory.
64-bit-native Operating systems tend to prefer `XDsdt` over `Dsdt`,
especially because there is no guarantee that the DSDT was allocated in
low memory addressable by only 32 bits.

This change shouldn't cause any issues; after all, the header signature
for FACP 6.4 is identical to that of FACP 2.0; apart from the new use of
`XDsdt`, this change is mostly modernization.