Fix duplicate definitions of gLoadedImage
Including a non-extern definition of `gLoadedImage` in FsHelpers.h
causes the compiler to emit a duplicate copy of `gLoadedImage` into
every translation unit that includes it. Since the only user of
`gLoadedImage` is _in FsHelpers.c_, it's reasonable to reduce the scope
of the variable.

If you'd like, we can reintroduce it as an `extern` in FsHelpers.h so
other parts of the code can access it.